Refactor to use cmake 3, build on FreeBSD
- Require CMake 3.2.0 or later.
- Drop schroot-common library; now dchroot is gone, it's only user
  is schroot, so merge with schroot from where it was originally
  split out
- Drop manual testing for C++11 features; use CMake compile features
  instead.
- Refactor CMake test logic and drop cleanup script.  The CMakeLists.txt
  for each test is now in the same directory rather than at the top level.
- Minor portability fixes for FreeBSD such as declaring environ in the
  run-parts tests.
