+++ bin/schroot/main.cc
@@ -18,9 +18,19 @@
 
 #include <config.h>
 
-#include <schroot/config.h>
+#include <bin-common/main.h>
 
-#include <schroot/main.h>
+#include <schroot/config.h>
+#ifdef SCHROOT_FEATURE_PAM
+#include <schroot/auth/pam.h>
+#include <schroot/auth/pam-conv.h>
+#include <schroot/auth/pam-conv-tty.h>
+#endif // SCHROOT_FEATURE_PAM
+#include <schroot/chroot/facet/factory.h>
+#include <schroot/keyfile-writer.h>
+#include <schroot/session.h>
+
+#include <bin/schroot/main.h>
 
 #include <cstdlib>
 #include <ctime>
@@ -35,23 +45,359 @@
 using std::endl;
 using boost::format;
 using schroot::_;
+using schroot::N_;
+
+namespace schroot
+{
+
+  template<>
+  error<bin::schroot::main::error_code>::map_type
+  error<bin::schroot::main::error_code>::error_strings =
+    {
+      // TRANSLATORS: %4% = file
+      {bin::schroot::main::CHROOT_FILE,       N_("No chroots are defined in ‘%4%’")},
+      // TRANSLATORS: %4% = file
+      // TRANSLATORS: %5% = file
+      {bin::schroot::main::CHROOT_FILE2,      N_("No chroots are defined in ‘%4%’ or ‘%5%’")},
+      // TRANSLATORS: %1% = file
+      {bin::schroot::main::CHROOT_NOTDEFINED, N_("The specified chroots are not defined in ‘%1%’")},
+      {bin::schroot::main::SESSION_INVALID,   N_("%1%: Invalid session name")}
+    };
+
+}
 
 namespace bin
 {
   namespace schroot
   {
 
-    main::main (schroot_common::options::ptr& options):
-      schroot_common::main("schroot",
-                           _("[OPTION…] [COMMAND] — run command or shell in a chroot"),
-                           options,
-                           true)
+    main::main (schroot::options::ptr& options):
+      bin::common::main("schroot",
+                        _("[OPTION…] [COMMAND] — run command or shell in a chroot"),
+                        options,
+                        true),
+      opts(options)
     {
     }
 
     main::~main ()
     {
     }
+
+    void
+    main::action_version (std::ostream& stream)
+    {
+      bin::common::main::action_version(stream);
+
+      stream << '\n'
+             << _("Installed chroot facets:")
+             << '\n';
+      ::schroot::chroot::facet::factory::print_facets(stream);
+
+      stream << std::flush;
+    }
+
+    void
+    main::action_info ()
+    {
+      for(auto chroot_name = this->chroot_names.begin();
+          chroot_name != this->chroot_names.end();
+          ++chroot_name)
+        {
+          // This should never fail, so no error handling here--we already
+          // validated everything when we got the chroot map.
+          ::schroot::chroot::config::chroot_map::const_iterator c = this->chroots.find(*chroot_name);
+          assert(c->second);
+
+          std::cout << c->second;
+          if (chroot_name + 1 != this->chroot_names.end())
+            std::cout << '\n';
+        }
+
+      std::cout << std::flush;
+    }
+
+    void
+    main::action_location ()
+    {
+      for(const auto& chroot_name : this->chroot_names)
+        {
+          // This should never fail, so no error handling here--we already
+          // validated everything when we got the chroot map.
+          const auto chroot = this->chroots.find(chroot_name);
+          assert(chroot->second);
+
+          std::cout << chroot->second->get_path() << '\n';
+        }
+
+      std::cout << std::flush;
+    }
+
+    void
+    main::action_config ()
+    {
+      std::cout << "# "
+        // TRANSLATORS: %1% = program name
+        // TRANSLATORS: %2% = program version
+        // TRANSLATORS: %3% = current date
+                << format(_("schroot configuration generated by %1% %2% on %3%"))
+        % this->program_name % VERSION % ::schroot::date(time(0))
+                << endl;
+      std::cout << endl;
+
+      ::schroot::keyfile info;
+
+      for(const auto& chroot_name : this->chroot_names)
+        {
+          // This should never fail, so no error handling here--we already
+          // validated everything when we got the chroot map.
+          const auto chroot = this->chroots.find(chroot_name);
+          assert(chroot->second);
+
+          // Generated chroots (e.g. source chroots) are not printed.
+          if (chroot->second->get_original())
+            info << chroot->second;
+        }
+
+      std::cout << ::schroot::keyfile_writer(info) << std::flush;
+    }
+
+    void
+    main::get_chroot_options ()
+    {
+
+      if (this->opts->all_chroots == true ||
+          this->opts->all_sessions == true ||
+          this->opts->all_source_chroots == true)
+        {
+          if (this->opts->all_chroots)
+            {
+              ::schroot::string_list chroots;
+              if (this->opts->action == options::ACTION_LIST &&
+                  !this->opts->exclude_aliases)
+                chroots = this->config->get_alias_list("chroot");
+              else
+                chroots = this->config->get_chroot_list("chroot");
+              this->chroot_names.insert(this->chroot_names.end(), chroots.begin(), chroots.end());
+            }
+          if (this->opts->all_sessions)
+            {
+              ::schroot::string_list sessions;
+              if (this->opts->action == options::ACTION_LIST &&
+                  !this->opts->exclude_aliases)
+                sessions = this->config->get_alias_list("session");
+              else
+                sessions = this->config->get_chroot_list("session");
+              this->chroot_names.insert(this->chroot_names.end(), sessions.begin(), sessions.end());
+            }
+          if (this->opts->all_source_chroots)
+            {
+              ::schroot::string_list sources;
+              if (this->opts->action == options::ACTION_LIST &&
+                  !this->opts->exclude_aliases)
+                sources = this->config->get_alias_list("source");
+              else
+                sources = this->config->get_chroot_list("source");
+              this->chroot_names.insert(this->chroot_names.end(), sources.begin(), sources.end());
+            }
+
+          // Validate and normalise
+          this->chroots = this->config->validate_chroots("", this->chroot_names);
+        }
+      else
+        {
+          // Search in the appropriate namespace.
+          std::string chroot_namespace("chroot");
+          if (this->opts->action == options::ACTION_SESSION_RECOVER ||
+              this->opts->action == options::ACTION_SESSION_RUN ||
+              this->opts->action == options::ACTION_SESSION_END)
+            chroot_namespace = "session";
+
+          // Validate and normalise
+          this->chroot_names = this->opts->chroots;
+          this->chroots = this->config->validate_chroots
+            (chroot_namespace, this->chroot_names);
+        }
+    }
+
+    void
+    main::load_config ()
+    {
+      this->config = ::schroot::chroot::config::ptr(new ::schroot::chroot::config);
+      /* The normal chroot list is used when starting a session or running
+         any chroot type or session, or displaying chroot information. */
+      if (this->opts->load_chroots == true)
+        {
+          this->config->add("chroot", SCHROOT_CONF);
+          this->config->add("chroot", SCHROOT_CONF_CHROOT_D);
+        }
+      /* The session chroot list is used when running or ending an
+         existing session, or displaying chroot information. */
+      if (this->opts->load_sessions == true)
+        this->config->add("session", SCHROOT_SESSION_DIR);
+    }
+
+    int
+    main::run_impl ()
+    {
+      if (this->opts->action == options::ACTION_HELP)
+        {
+          action_help(std::cout);
+          return EXIT_SUCCESS;
+        }
+
+      if (this->opts->action == options::ACTION_VERSION)
+        {
+          action_version(std::cout);
+          return EXIT_SUCCESS;
+        }
+
+      /* Initialise chroot configuration. */
+      load_config();
+
+      if (this->opts->load_chroots &&
+          this->config->get_chroots("chroot").empty() &&
+          this->opts->quiet == false)
+        log_exception_warning(error(CHROOT_FILE2, SCHROOT_CONF, SCHROOT_CONF_CHROOT_D));
+
+      /* Get list of chroots to use */
+      get_chroot_options();
+      if (this->chroots.empty())
+        {
+          if (!(this->opts->all_chroots == true ||
+                this->opts->all_sessions == true ||
+                this->opts->all_source_chroots == true))
+            throw error(SCHROOT_CONF, CHROOT_NOTDEFINED);
+          else
+            {
+              // If one of the --all options was used, then don't treat
+              // the lack of chroots as an error.  TODO: Also check if any
+              // additional chroots were specified with -c; this needs
+              // changes in get_chroot_options.
+              if (this->opts->verbose)
+                log_exception_warning(error((this->opts->all_chroots == true) ?
+                                            SCHROOT_CONF : SCHROOT_SESSION_DIR,
+                                            CHROOT_NOTDEFINED));
+              return EXIT_SUCCESS;
+            }
+        }
+      this->chroot_objects.clear();
+      for(const auto& chroot_name : this->chroot_names)
+        {
+          // This should never fail, so no error handling here--we already
+          // validated everything when we got the chroot map.
+          const auto chroot = this->chroots.find(chroot_name);
+          assert(chroot->second);
+          ::schroot::session::chroot_list::value_type e;
+          e.alias = chroot->first;
+          e.chroot = chroot->second;
+          chroot_objects.push_back(e);
+        }
+
+      /* Print chroot list. */
+      if (this->opts->action == options::ACTION_LIST)
+        {
+          action_list();
+          return EXIT_SUCCESS;
+        }
+
+      if (this->config->find_alias("session", this->opts->session_name))
+        throw error(this->opts->session_name, SESSION_INVALID);
+
+      /* Print chroot information for specified chroots. */
+      if (this->opts->action == options::ACTION_INFO)
+        {
+          action_info();
+          return EXIT_SUCCESS;
+        }
+      if (this->opts->action == options::ACTION_LOCATION)
+        {
+          action_location();
+          return EXIT_SUCCESS;
+        }
+      if (this->opts->action == options::ACTION_CONFIG)
+        {
+          action_config();
+          return EXIT_SUCCESS;
+        }
+
+      /* Create a session. */
+      ::schroot::session::operation sess_op(::schroot::session::OPERATION_AUTOMATIC);
+      if (this->opts->action == options::ACTION_SESSION_BEGIN)
+        sess_op = ::schroot::session::OPERATION_BEGIN;
+      else if (this->opts->action == options::ACTION_SESSION_RECOVER)
+        sess_op = ::schroot::session::OPERATION_RECOVER;
+      else if (this->opts->action == options::ACTION_SESSION_RUN)
+        sess_op = ::schroot::session::OPERATION_RUN;
+      else if (this->opts->action == options::ACTION_SESSION_END)
+        sess_op = ::schroot::session::OPERATION_END;
+
+      try
+        {
+          create_session(sess_op);
+          add_session_auth();
+
+          if (!this->opts->command.empty())
+            this->session->get_auth()->set_command(this->opts->command);
+          if (!this->opts->directory.empty())
+            this->session->get_auth()->set_wd(this->opts->directory);
+          if (!this->opts->shell.empty())
+            this->session->set_shell_override(this->opts->shell);
+          this->session->set_preserve_environment(this->opts->preserve);
+          this->session->set_session_id(this->opts->session_name);
+          this->session->set_force(this->opts->session_force);
+          if (this->opts->quiet)
+            this->session->set_verbosity("quiet");
+          else if (this->opts->verbose)
+            this->session->set_verbosity("verbose");
+          this->session->set_user_options(this->opts->useroptions_map);
+
+          /* Run session. */
+          this->session->run();
+        }
+      catch (const std::runtime_error& e)
+        {
+          if (!this->opts->quiet)
+            ::schroot::log_exception_error(e);
+        }
+
+      if (this->session)
+        return this->session->get_child_status();
+      else
+        return EXIT_FAILURE;
+    }
+
+    void
+    main::add_session_auth ()
+    {
+      // Add PAM authentication handler.  If PAM isn't available, just
+      // continue to use the default handler
+
+#ifdef SCHROOT_FEATURE_PAM
+      ::schroot::auth::auth::ptr auth = ::schroot::auth::pam::create("schroot");
+
+      ::schroot::auth::pam_conv::auth_ptr pam_auth =
+        std::dynamic_pointer_cast<::schroot::auth::pam>(auth);
+
+      ::schroot::auth::pam_conv::ptr conv = ::schroot::auth::pam_conv_tty::create(pam_auth);
+
+
+      /* Set up authentication timeouts. */
+      time_t curtime = 0;
+      time(&curtime);
+      conv->set_warning_timeout(curtime + 15);
+      conv->set_fatal_timeout(curtime + 20);
+      pam_auth->set_conv(conv);
+
+      this->session->set_auth(auth);
+#endif // SCHROOT_FEATURE_PAM
+
+      if (!this->opts->user.empty())
+        this->session->get_auth()->set_user(this->opts->user);
+    }
+
+
+
 
     void
     main::action_list ()
@@ -69,15 +415,6 @@ namespace bin
 
       this->session = ::schroot::session::ptr
         (new ::schroot::session("schroot", sess_op, this->chroot_objects));
-    }
-
-    void
-    main::add_session_auth ()
-    {
-      schroot_common::main::add_session_auth();
-
-      if (!this->options->user.empty())
-        this->session->get_auth()->set_user(this->options->user);
     }
 
   }
