--- lib/schroot-common/CMakeLists.txt
@@ -1,26 +1,0 @@
-# Copyright © 2004-2013  Roger Leigh <rleigh@codelibre.net>
-#
-# schroot is free software: you can redistribute it and/or modify it
-# under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# schroot is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-# General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see
-# <http://www.gnu.org/licenses/>.
-#
-#####################################################################
-
-set(schroot_common_sources
-    options.h
-    options.cc
-    main.h
-    main.cc)
-
-add_library(schroot-common STATIC ${schroot_common_sources})
-target_link_libraries(schroot-common bin-common)
