--- test/cleanup-test-data
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-set -e
-
-rm -rf test/testdata
