+++ test/setup-test-data
@@ -7,4 +7,4 @@ rm -rf test/testdata
 mkdir test/testdata
 cp -r ${srcdir}/test/*.ex* test/testdata
 mkdir test/testdata/run-parts.ex2
-chown -R root:root test/testdata
+chown -R root test/testdata
